Crud api

+++ web-search/backend/collections/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class Collection(models.Model):
     # big int 
-    id = models.IntegerField() 
+    id = models.BigIntegerField()
     name = models.CharField()
     notes = models.CharField()
     created_at = models.DateTimeField(auto_now_add= True)
